Refactor local import analysis and add unit tests
