+++ domain/local_import/media_file.py
@@ -3,9 +3,9 @@
 from __future__ import annotations
 
 import os
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from datetime import datetime, timezone
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Optional, Protocol
 
 from core.utils import get_file_date_from_exif, get_file_date_from_name
 
@@ -46,53 +46,115 @@ class MediaFileAnalysis:
     relative_path: str
 
 
+class MediaMetadataProvider(Protocol):
+    """メディアファイル解析に必要なメタデータ取得の抽象化."""
+
+    def calculate_file_hash(self, file_path: str) -> str:
+        ...
+
+    def extract_exif_data(self, file_path: str) -> Dict[str, Any]:
+        ...
+
+    def extract_video_metadata(self, file_path: str) -> Dict[str, Any]:
+        ...
+
+    def generate_filename(self, shot_at: datetime, extension: str, file_hash: str) -> str:
+        ...
+
+    def get_image_dimensions(self, file_path: str) -> tuple[Optional[int], Optional[int], Optional[int]]:
+        ...
+
+    def get_relative_path(self, shot_at: datetime, destination_filename: str) -> str:
+        ...
+
+
+@dataclass(frozen=True)
+class DefaultMediaMetadataProvider:
+    """ドメインのメタデータ関数を利用するデフォルト実装."""
+
+    def calculate_file_hash(self, file_path: str) -> str:
+        return calculate_file_hash(file_path)
+
+    def extract_exif_data(self, file_path: str) -> Dict[str, Any]:
+        return extract_exif_data(file_path)
+
+    def extract_video_metadata(self, file_path: str) -> Dict[str, Any]:
+        return extract_video_metadata(file_path)
+
+    def generate_filename(self, shot_at: datetime, extension: str, file_hash: str) -> str:
+        return generate_filename(shot_at, extension, file_hash)
+
+    def get_image_dimensions(self, file_path: str) -> tuple[Optional[int], Optional[int], Optional[int]]:
+        return get_image_dimensions(file_path)
+
+    def get_relative_path(self, shot_at: datetime, destination_filename: str) -> str:
+        return get_relative_path(shot_at, destination_filename)
+
+
+@dataclass(frozen=True)
+class MediaFileAnalyzer:
+    """メディアファイルの解析ロジックを担当するドメインサービス."""
+
+    metadata_provider: MediaMetadataProvider = field(
+        default_factory=DefaultMediaMetadataProvider
+    )
+
+    def analyze(self, file_path: str) -> MediaFileAnalysis:
+        source = ImportFile(file_path)
+        extension = source.extension
+        file_size = os.path.getsize(file_path)
+        file_hash = self.metadata_provider.calculate_file_hash(file_path)
+        is_video = extension in SUPPORTED_VIDEO_EXTENSIONS
+        mime_type = MIME_TYPE_BY_EXTENSION.get(extension, DEFAULT_MIME_TYPE)
+
+        width: Optional[int] = None
+        height: Optional[int] = None
+        orientation: Optional[int] = None
+        duration_ms: Optional[int] = None
+        exif_data: Dict[str, Any] = {}
+        video_metadata: Dict[str, Any] = {}
+
+        if is_video:
+            video_metadata = self.metadata_provider.extract_video_metadata(file_path)
+            width = video_metadata.get("width")
+            height = video_metadata.get("height")
+            duration_ms = video_metadata.get("duration_ms")
+        else:
+            width, height, orientation = self.metadata_provider.get_image_dimensions(file_path)
+            exif_data = self.metadata_provider.extract_exif_data(file_path)
+
+        shot_at = _resolve_shot_at(source, exif_data, video_metadata)
+        destination_filename = self.metadata_provider.generate_filename(
+            shot_at, extension, file_hash
+        )
+        relative_path = self.metadata_provider.get_relative_path(
+            shot_at, destination_filename
+        )
+
+        return MediaFileAnalysis(
+            source=source,
+            extension=extension,
+            file_size=file_size,
+            file_hash=file_hash,
+            mime_type=mime_type,
+            is_video=is_video,
+            width=width,
+            height=height,
+            duration_ms=duration_ms,
+            orientation=orientation,
+            shot_at=shot_at,
+            exif_data=exif_data,
+            video_metadata=video_metadata,
+            destination_filename=destination_filename,
+            relative_path=relative_path,
+        )
+
+
 def analyze_media_file(file_path: str) -> MediaFileAnalysis:
     """ファイルを解析して :class:`MediaFileAnalysis` を返す。"""
 
-    source = ImportFile(file_path)
-    extension = source.extension
-    file_size = os.path.getsize(file_path)
-    file_hash = calculate_file_hash(file_path)
-    is_video = extension in SUPPORTED_VIDEO_EXTENSIONS
-    mime_type = MIME_TYPE_BY_EXTENSION.get(extension, DEFAULT_MIME_TYPE)
-
-    width: Optional[int] = None
-    height: Optional[int] = None
-    orientation: Optional[int] = None
-    duration_ms: Optional[int] = None
-    exif_data: Dict[str, Any] = {}
-    video_metadata: Dict[str, Any] = {}
-
-    if is_video:
-        video_metadata = extract_video_metadata(file_path)
-        width = video_metadata.get("width")
-        height = video_metadata.get("height")
-        duration_ms = video_metadata.get("duration_ms")
-    else:
-        width, height, orientation = get_image_dimensions(file_path)
-        exif_data = extract_exif_data(file_path)
-
-    shot_at = _resolve_shot_at(source, exif_data, video_metadata)
-    destination_filename = generate_filename(shot_at, extension, file_hash)
-    relative_path = get_relative_path(shot_at, destination_filename)
-
-    return MediaFileAnalysis(
-        source=source,
-        extension=extension,
-        file_size=file_size,
-        file_hash=file_hash,
-        mime_type=mime_type,
-        is_video=is_video,
-        width=width,
-        height=height,
-        duration_ms=duration_ms,
-        orientation=orientation,
-        shot_at=shot_at,
-        exif_data=exif_data,
-        video_metadata=video_metadata,
-        destination_filename=destination_filename,
-        relative_path=relative_path,
-    )
+    analyzer = MediaFileAnalyzer()
+    return analyzer.analyze(file_path)
 
 
 def _resolve_shot_at(
@@ -114,5 +176,11 @@ def _resolve_shot_at(
     return shot_at
 
 
-__all__ = ["MediaFileAnalysis", "analyze_media_file"]
+__all__ = [
+    "DefaultMediaMetadataProvider",
+    "MediaFileAnalysis",
+    "MediaFileAnalyzer",
+    "MediaMetadataProvider",
+    "analyze_media_file",
+]
