+++ tests/conftest.py
@@ -83,6 +83,7 @@ ALWAYS_RUN = {
     "tests/test_local_import_duplicate_refresh.py",
     "tests/test_local_import.py",
     "tests/test_local_import_ui.py",
+    "tests/test_local_import_services.py",
 }
